moved ssn validator to new file and allow only one format
