+++ src/registrationSystem/utils.py
@@ -1,4 +1,3 @@
-from django.core import validators
 from django.template.loader import render_to_string
 from django.core.mail import EmailMessage
 from django.core.exceptions import FieldError
@@ -95,27 +94,3 @@ def is_utn_member(person_nr):
     )
 
     return r.json()['is_member']
-
-
-# TODO: Change regex to only support YYYYMMDD-XXXX
-class SSNValidator(validators.RegexValidator):
-    century = r'[1-2][0|9]'  # YY-- Only allows 19-- and 20--
-    decade = r'[0-9]{2}'  # --YY
-    month = r'[0-1][0-9]'
-    day = r'[0-3][0-9]'
-    last_four = r'(T|t|[0-9])[0-9]{3}'  # Allows T-number SSN
-
-    def __init__(self):
-        regex_pattern = \
-            r'^(' + self.century + r')?' + \
-            self.decade + \
-            self.month + \
-            self.day + \
-            r'(-|\+)?' + \
-            self.last_four + \
-            r'$'
-
-        super(SSNValidator, self).__init__(
-            regex=regex_pattern,
-            message='Use the format YYYYMMDD-XXXX.'
-        )
